Release 0.2.0
Summary of changes:

* Bump support for Go to 1.24+
* Change all functions that return `Builder` to return `*Builder` pointer instead
* Add built-in support for [zap](https://github.com/uber-go/zap) logging
* Fix issue causing `Translator` to be ignored when specified
* Improve documentation
* Bump all dependencies to latest versions
