+++ .github/workflows/ci.yml
@@ -20,8 +20,8 @@ jobs:
     strategy:
       matrix:
         go-version:
-          - '1.21.x'
-          - '1.22.x'
+          - '1.24.x'
+          - '1.25.x'
 
     steps:
       - name: Checkout
