+++ go.mod
@@ -1,17 +1,17 @@
 module github.com/neocotic/go-problem
 
-go 1.21
+go 1.24
 
 require (
 	github.com/google/uuid v1.6.0
 	github.com/neocotic/go-optional v0.1.2
-	github.com/stretchr/testify v1.9.0
+	github.com/stretchr/testify v1.11.1
+	go.uber.org/zap v1.27.0
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	go.uber.org/multierr v1.10.0 // indirect
-	go.uber.org/zap v1.27.0 // indirect
+	go.uber.org/multierr v1.11.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
